pip_requirements.txt
django=3.2.8
djangorestframework=3.12.4
djoser=2.1.0